+++ tests/testthat/test-sampling.R
@@ -1,4 +1,4 @@
-context("Sampling")
+context("Function test for sampling()")
 
 # jfa version 0.1.0